playing_with_C-1 Add Safe Create function
added a way to create and manage new objects into memory, with proper disposal.

requirements for a new dynamic struct is that it contains a MemorySafe object, an AddXToMemory function that takes in a MemoryHolder pointer and adds all containing variables to it as well as a Create function that sets the MemorySafe.AddToMemory function to call the AddXToMemory function when needed.

a parent object should call the child objects' MemorySafe.AddToMemory functions